feat(modes): make modes available for specific buffers
Users can pass value of buffer or any other options to modes.toggleMode
funtion to make it available to specific buffer.
If toggling mode globally after it is already enabled locally it will disabled from
buffer and enabled globally. NOTE: This functionality seems to be flaky
need to updated later.